21/269. Submitting the payment

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -67,9 +67,26 @@ export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
     const orderToCreate = this.getOrderToCreate(basket);
     this.checkoutService.creatOrder(orderToCreate).subscribe((order: IOrder) => {
       this.toastr.success('Order created successfully');
-      this.basketService.deleteLocalBasket(basket.id);
-      const navigationExtras: NavigationExtras = {state: order};
-      this.router.navigate(['checkout/success'], navigationExtras);
+      this.stripe.confirmCardPayment(basket.clientSecret, {
+        payment_method: {
+          card: this.cardNumber,
+          billing_details: {
+            name: this.checkoutForm.get('paymentForm').get('nameOnCard').value
+          }
+        }
+      }).then(result => {
+        console.log(result);
+        if (result.paymentIntent) {
+         this.basketService.deleteLocalBasket(basket.id);
+         const navigationExtras: NavigationExtras = {state: order};
+         this.router.navigate(['checkout/success'], navigationExtras);
+
+        } else {
+          this.toastr.error('Payment error');
+
+        }
+      })
+   
     }, error => {
       this.toastr.error(error.message);
       console.log(error);
